% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{parse_url}
\alias{parse_url}
\title{Parse a vector of URLs and return a DataFrame with parsed components}
\usage{
parse_url(urls)
}
\description{
Parse a vector of URLs and return a DataFrame with parsed components
}
