% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{rs_url_parse}
\alias{rs_url_parse}
\title{Use url crate to parse a url}
\usage{
rs_url_parse(url)
}
\arguments{
\item{url}{character. one or more URL to be parsed}
}
\value{
A data frame of the url components:
href, protocol, username, password, host, hostname, port, pathname, search, and hash
}
\description{
Use url crate to parse a url
}
\examples{
rs_url_parse("https://user_1:password_1@example.org:8080/dir/../api?q=1#frag")
}
