% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{url_parse}
\alias{url_parse}
\title{Parse a vector of URLs and return a DataFrame with parsed components}
\usage{
url_parse(urls)
}
\description{
Parse a vector of URLs and return a DataFrame with parsed components
}
